Add files via upload
The objective of this dataset was to analyze the Big Data in apache pig. We had 
worked on the HDFS cluster and written many Pig scripts using different 
functions, mainly String, DateTime, Eval, Math and Tuple and various 
operators. The operating system which was used is Ubuntu. The purpose of 
analyzing this dataset was to fetch our requirements easily with less processing 
time.